push notification badge count issue solved
