+++ app/src/main/java/com/qatarmuseums/qatarmuseumsapp/services/MyFirebaseMessagingService.java
@@ -124,6 +124,8 @@ public class MyFirebaseMessagingService extends FirebaseMessagingService {
 
     private void sendRegistrationToServer(String token) {
         // TODO: Implement this method to send token to your app server.
+
+
     }
 
     private void handleNow(String s) {
